Refactor Answer and restructure to folders
