+++ src/App.js
@@ -1,0 +1,15 @@
+import { Component } from 'react';
+import Calculator from './Calculator/Calculator';
+
+class App extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (<Calculator />);
+  }
+}
+
+export default App;
