+++ src/Calculator/Calculator.css
@@ -1,0 +1,21 @@
+* {
+  box-sizing: border-box;
+}
+
+.inline {
+  display: inline;
+  
+ 
+}
+
+.centerBox {
+  display: flex;
+  flex-direction: column;
+ justify-content: center;
+ align-items: center;
+}
+
+.flex {
+  display: flex;
+
+}
