--- src/components/Answer.js
@@ -1,5 +1,0 @@
-import './Answer.css';
-
-const Answer = () => <div className="answer">0</div>;
-
-export default Answer;
