+++ src/components/Answer/Answer.js
@@ -1,0 +1,15 @@
+import { Component } from 'react';
+import './Answer.css';
+
+class Answer extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (<div className="answer">0</div>);
+  }
+}
+
+export default Answer;
