--- src/components/Calculator.css
@@ -1,14 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-.inline {
-  display: inline;
-  
- 
-}
-
-.flex {
-  display: flex;
-
-}
