+++ src/components/Tiles/Tile.css
@@ -10,7 +10,7 @@ ul {
   justify-content: center;
   text-align: center;
   width: 70px;
-  height: 70px;
+  height: 50px;
   font-size: 0.9rem;
 }
