+++ src/components/Tiles/Tile.js
@@ -1,0 +1,7 @@
+import './Tile.css';
+
+export { default as OpperationTile } from './OpperationTile';
+
+export { default as WideTile } from './WideTile';
+
+export { default as NumberTile } from './NumberTile';
