+++ src/components/lists/NumberList.js
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import { NumberTile } from '../Tile';
+import { NumberTile } from '../Tiles/Tile';
 
 class NumberList extends Component {
   constructor(props) {
@@ -12,7 +12,7 @@ class NumberList extends Component {
       <ul className="num-list">
 
         {
-      Array.from(Array(9).keys()).map((i) => <NumberTile key={i} value={i + 1} />)
+      Array.from(Array(9).keys()).map((i) => <NumberTile key={i} value={`${i + 1}`} />)
     }
       </ul>
     );
